+++ crates/consensus/beacon/Cargo.toml
@@ -20,6 +20,7 @@ reth-errors.workspace = true
 reth-provider.workspace = true
 reth-tasks.workspace = true
 reth-payload-builder.workspace = true
+reth-payload-builder-primitives.workspace = true
 reth-payload-primitives.workspace = true
 reth-payload-validator.workspace = true
 reth-prune.workspace = true
