+++ crates/e2e-test-utils/Cargo.toml
@@ -17,6 +17,7 @@ reth-tracing.workspace = true
 reth-db = { workspace = true, features = ["test-utils"] }
 reth-rpc-layer.workspace = true
 reth-payload-builder = { workspace = true, features = ["test-utils"] }
+reth-payload-builder-primitives.workspace = true
 reth-payload-primitives.workspace = true
 reth-provider.workspace = true
 reth-node-builder = { workspace = true, features = ["test-utils"] }
