+++ crates/engine/local/Cargo.toml
@@ -19,6 +19,7 @@ reth-engine-tree.workspace = true
 reth-evm.workspace = true
 reth-ethereum-engine-primitives.workspace = true
 reth-payload-builder.workspace = true
+reth-payload-builder-primitives.workspace = true
 reth-payload-primitives.workspace = true
 reth-payload-validator.workspace = true
 reth-provider.workspace = true
