+++ crates/node/api/Cargo.toml
@@ -18,6 +18,7 @@ reth-evm.workspace = true
 reth-provider.workspace = true
 reth-engine-primitives.workspace = true
 reth-transaction-pool.workspace = true
+reth-payload-builder-primitives.workspace = true
 reth-payload-primitives.workspace = true
 reth-tasks.workspace = true
 reth-network-api.workspace = true
