+++ crates/payload/basic/Cargo.toml
@@ -19,6 +19,7 @@ reth-primitives-traits.workspace = true
 reth-transaction-pool.workspace = true
 reth-provider.workspace = true
 reth-payload-builder.workspace = true
+reth-payload-builder-primitives.workspace = true
 reth-payload-primitives.workspace = true
 reth-tasks.workspace = true
 reth-evm.workspace = true
