+++ crates/payload/builder-primitives/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "reth-payload-primitives"
+name = "reth-payload-builder-primitives"
 version.workspace = true
 edition.workspace = true
 rust-version.workspace = true
@@ -13,19 +13,21 @@ workspace = true
 
 [dependencies]
 # reth
-reth-chainspec.workspace = true
-reth-primitives.workspace = true
-reth-chain-state.workspace = true
+reth-errors.workspace = true
+reth-payload-primitives.workspace = true
+reth-transaction-pool.workspace = true
+revm-primitives.workspace = true
 
 # alloy
-alloy-eips.workspace = true
 alloy-primitives.workspace = true
 alloy-rpc-types-engine = { workspace = true, features = ["serde"] }
-op-alloy-rpc-types-engine = { workspace = true, optional = true }
+
+# async
+async-trait.workspace = true
+pin-project.workspace = true
+tokio = { workspace = true, features = ["sync"] }
+tokio-stream.workspace = true
 
 # misc
-serde.workspace = true
 thiserror.workspace = true
-
-[features]
-op = ["dep:op-alloy-rpc-types-engine"]
+tracing.workspace = true
