+++ crates/payload/builder/Cargo.toml
@@ -15,6 +15,7 @@ workspace = true
 # reth
 reth-primitives = { workspace = true, optional = true }
 reth-chain-state.workspace = true
+reth-payload-builder-primitives.workspace = true
 reth-payload-primitives.workspace = true
 reth-ethereum-engine-primitives.workspace = true
