+++ crates/payload/builder/src/test_utils.rs
@@ -7,7 +7,8 @@ use crate::{
 
 use alloy_primitives::U256;
 use reth_chain_state::{CanonStateNotification, ExecutedBlock};
-use reth_payload_primitives::{PayloadBuilderError, PayloadKind, PayloadTypes};
+use reth_payload_builder_primitives::PayloadBuilderError;
+use reth_payload_primitives::{PayloadKind, PayloadTypes};
 use reth_primitives::Block;
 use std::{
     future::Future,
