+++ crates/rpc/rpc-engine-api/src/error.rs
@@ -4,7 +4,8 @@ use jsonrpsee_types::error::{
 };
 use reth_beacon_consensus::BeaconForkChoiceUpdateError;
 use reth_engine_primitives::BeaconOnNewPayloadError;
-use reth_payload_primitives::{EngineObjectValidationError, PayloadBuilderError};
+use reth_payload_builder_primitives::PayloadBuilderError;
+use reth_payload_primitives::EngineObjectValidationError;
 use thiserror::Error;
 
 /// The Engine API result type
